+++ src/Fauna/reproduction_models.test.cpp
@@ -8,8 +8,7 @@
 #include "reproduction_models.h"
 using namespace Fauna;
 
-
-TEST_CASE("Fauna::ReprIlliusOconnor2000", "") {
+TEST_CASE("Fauna::ReproductionLogistic", "") {
   // NOTE: We are more tolerant with the Approx() function of
   // the CATCH framework (by adjusting Approx().epsilon().
 
@@ -22,8 +21,8 @@ TEST_CASE("Fauna::ReprIlliusOconnor2000", "") {
 
   // exceptions
   SECTION("exceptions") {
-    CHECK_THROWS(ReprIlliusOconnor2000(season, -1.0));
-    ReprIlliusOconnor2000 rep(season, INC);
+    CHECK_THROWS(ReproductionLogistic(season, -1.0));
+    ReproductionLogistic rep(season, INC);
     CHECK_THROWS(rep.get_offspring_density(-1, OPT));
     CHECK_THROWS(rep.get_offspring_density(365, OPT));
     CHECK_THROWS(rep.get_offspring_density(START, -0.1));
@@ -33,8 +32,8 @@ TEST_CASE("Fauna::ReprIlliusOconnor2000", "") {
   SECTION("higher annual increase makes more offspring") {
     const double INC2 = INC * 1.5;
     REQUIRE(INC2 > INC);
-    ReprIlliusOconnor2000 rep1(season, INC);
-    ReprIlliusOconnor2000 rep2(season, INC2);
+    ReproductionLogistic rep1(season, INC);
+    ReproductionLogistic rep2(season, INC2);
     CHECK(rep1.get_offspring_density(START, OPT) <
           rep2.get_offspring_density(START, OPT));
     CHECK(rep1.get_offspring_density(START, OPT) < INC);
@@ -43,7 +42,7 @@ TEST_CASE("Fauna::ReprIlliusOconnor2000", "") {
 
   SECTION("better body condition makes more offspring") {
     const double BAD = OPT / 2.0;  // bad body condition
-    ReprIlliusOconnor2000 rep(season, INC);
+    ReproductionLogistic rep(season, INC);
     CHECK(rep.get_offspring_density(START, BAD) <
           rep.get_offspring_density(START, OPT));
   }
@@ -51,7 +50,7 @@ TEST_CASE("Fauna::ReprIlliusOconnor2000", "") {
   SECTION("one-day season length -> all offspring at once") {
     const double BAD = OPT / 2.0;  // bad body condition
     BreedingSeason season_short(START, 1);
-    ReprIlliusOconnor2000 rep(season_short, INC);
+    ReproductionLogistic rep(season_short, INC);
     CHECK(rep.get_offspring_density(START, OPT) == Approx(INC).epsilon(0.05));
     CHECK(rep.get_offspring_density(START, BAD) < INC);
 
@@ -62,7 +61,7 @@ TEST_CASE("Fauna::ReprIlliusOconnor2000", "") {
   }
 
   SECTION("Sum of offspring over year must be max. annual increase") {
-    ReprIlliusOconnor2000 rep(season, INC);
+    ReproductionLogistic rep(season, INC);
     // sum over whole year
     double sum_year = 0.0;
     for (int d = 0; d < 365; d++)
@@ -76,4 +75,3 @@ TEST_CASE("Fauna::ReprIlliusOconnor2000", "") {
     CHECK(sum_season == Approx(INC).epsilon(0.05));
   }
 }
-